Working example of accessing the themoviedb.org API with Retrofit
On clicking the dummy menu item, the /movie/now_playing API is
accessed and the lists of movies contained in the response is
loaded into a MovieList (List<Movie>) object by Retrofit.

This List<Movie> can then be the data source for the ArrayAdapter
of the ListView on the MainActivity.