Working solution with dummy trailers
Adding a list of trailers in YouTubePlayerFragments or
YouTubePlayerViews is more complicated than exptected. This version
therefore displays only the link of each trailer of a movie. It
should be reworked to contain a YouTube player for each trailer
(or a thumbnail image that is replaced by a YouTubePlayerFragment
when clicked).