Integrate Stetho for network inspection
Stetho is a Chrome-based debugging tool for Android.

Stetho network inspection is enabled for the OkHttp client used by
Retrofit and Glide (Glide is set up to use an OkHttp client).

Now, when the device is connected and the app is running, just
type chrome://inspect in the address bar of Chrome and then the
app appears with an "inspect" link  at the bottom of the Devices
page.