Add click-wrap dialog at first run of app
The dialog displays the terms of use (which are saved as an HTML
asset) in a WebView and requires the user to accept them by checking
a checkbox in order to continue. If the user dismisses the dialog
without accepting the terms of use, the app terminates.

The same terms of use are also displayed in a WebView in the
TermsActivity, which can be opened at any time from the MainActivity
via an options menu item.